Move FSMSnapshot to own source file
